projections: pass telemetry through the API and small cleanups
