+++ packages/server/src/routes/projections.ts
@@ -1,9 +1,4 @@
-import type {
-  OntologySource,
-  ProjectionDefinition,
-  Sixb,
-  TelemetryProjectionDefinition,
-} from "@sixb/core"
+import type { OntologySource, Sixb } from "@sixb/core"
 import type { Elysia } from "elysia"
 import { ErrorResponseSchema } from "../schemas/common"
 import {
@@ -20,7 +15,7 @@ export function registerProjectionRoutes(app: Elysia, sixb: Sixb<readonly Ontolo
         return {
           objectProjections: [...sixb.getObjectProjections()],
           linkProjections: [...sixb.getLinkProjections()],
-          telemetryProjections: sixb.getTelemetryProjections().map(serializeTelemetryProjection),
+          telemetryProjections: [...sixb.getTelemetryProjections()],
         }
       },
       {
@@ -45,7 +40,7 @@ export function registerProjectionRoutes(app: Elysia, sixb: Sixb<readonly Ontolo
           set.status = 404
           return { error: `Projection '${params.projectionId}' not found` }
         }
-        return serializeProjection(found)
+        return found
       },
       {
         params: ProjectionParamsSchema,
@@ -57,26 +52,4 @@ export function registerProjectionRoutes(app: Elysia, sixb: Sixb<readonly Ontolo
         },
       }
     )
-}
-
-function serializeProjection(projection: ProjectionDefinition) {
-  if (projection._tag !== "TelemetryProjectionDefinition") {
-    return projection
-  }
-
-  return serializeTelemetryProjection(projection)
-}
-
-function serializeTelemetryProjection(projection: TelemetryProjectionDefinition) {
-  return {
-    _tag: projection._tag,
-    id: projection.id,
-    objectTypeId: projection.objectTypeId,
-    propertyId: projection.propertyId,
-    datasetId: projection.datasetId,
-    objectIdField: projection.objectIdField,
-    atField: projection.atField,
-    valueField: projection.valueField,
-    ...(projection.unitField !== undefined ? { unitField: projection.unitField } : {}),
-  }
 }
